+++ src/main/java/de/pt400c/defaultsettings/gui/ScrollbarSegment.java
@@ -14,7 +14,7 @@ public class ScrollbarSegment extends ButtonSegment {
 	
 	private final ScrollableSegment superScrollable;
 	
-	private double distanceY = 0;
+	private float distanceY = 0;
 
 	public ScrollbarSegment(GuiScreen gui, float posX, float posY, int width, int height, ScrollableSegment segment) {
 		super(gui, posX, posY, null, null, width, height, 0);
@@ -22,7 +22,7 @@ public class ScrollbarSegment extends ButtonSegment {
 	}
 
 	@Override
-	public void render(float mouseX, float mouseY, float partialTicks) {
+	public void render(int mouseX, int mouseY, float partialTicks) {
 
 		if (this.grabbed) {
 			final int factor = (int) (mouseY - distanceY);
@@ -35,8 +35,8 @@ public class ScrollbarSegment extends ButtonSegment {
 			if (this.posY > this.superScrollable.posY + this.superScrollable.height - height)
 				this.posY = this.superScrollable.posY + this.superScrollable.height - height;
 
-			final double distance = Math.round(this.superScrollable.height - height);
-			final double pos = Math.round(this.posY - this.superScrollable.posY);
+			final float distance = Math.round(this.superScrollable.height - height);
+			final float pos = Math.round(this.posY - this.superScrollable.posY);
 
 			final float tempAdd = (int) (this.superScrollable.getPosY() + this.superScrollable.height - 1 - 20 * (this.superScrollable.list.size() - 1) - this.superScrollable.getPosY() - 18);
 
@@ -65,16 +65,15 @@ public class ScrollbarSegment extends ButtonSegment {
 
 		drawRect(this.getPosX(), this.getPosY(), this.getPosX() + this.getWidth(), this.getPosY() + this.getHeight(), 0xffe0e0e0, true, null, false);
 
-		drawRect(this.getPosX() + this.width / 2 - 2D, this.getPosY() + this.height / 2 - 3, this.getPosX() + this.width / 2 + 2D, this.getPosY() + this.height / 2 - 3 + 1, 0xff373737, true, null, false);
+		drawRect(this.getPosX() + this.width / 2 - 2F, this.getPosY() + this.height / 2 - 3, this.getPosX() + this.width / 2 + 2F, this.getPosY() + this.height / 2 - 3 + 1, 0xff373737, true, null, false);
 
-		drawRect(this.getPosX() + this.width / 2 - 2D, this.getPosY() + this.height / 2, this.getPosX() + this.width / 2 + 2D, this.getPosY() + this.height / 2 + 1, 0xff373737, true, null, false);
+		drawRect(this.getPosX() + this.width / 2 - 2F, this.getPosY() + this.height / 2, this.getPosX() + this.width / 2 + 2F, this.getPosY() + this.height / 2 + 1, 0xff373737, true, null, false);
 
-		drawRect(this.getPosX() + this.width / 2 - 2D, this.getPosY() + this.height / 2 + 3, this.getPosX() + this.width / 2 + 2D, this.getPosY() + this.height / 2 + 1 + 3, 0xff373737, true, null, false);
-
+		drawRect(this.getPosX() + this.width / 2 - 2F, this.getPosY() + this.height / 2 + 3, this.getPosX() + this.width / 2 + 2F, this.getPosY() + this.height / 2 + 1 + 3, 0xff373737, true, null, false);
 	}
 	
 	@Override
-	public boolean mouseClicked(double mouseX, double mouseY, int button) {
+	public boolean mouseClicked(int mouseX, int mouseY, int button) {
 
 		if (this.isSelected(mouseX, mouseY)) {
 			this.grabbed = true;
@@ -88,16 +87,15 @@ public class ScrollbarSegment extends ButtonSegment {
 	}
 
 	@Override
-	public boolean mouseDragged(double mouseX, double mouseY, int button) {
+	public boolean mouseDragged(int mouseX, int mouseY, int button) {
 		return super.mouseDragged(mouseX, mouseY, button);
 	}
 
 	@Override
-	public boolean mouseReleased(double mouseX, double mouseY, int button) {
+	public boolean mouseReleased(int mouseX, int mouseY, int button) {
 		if (this.grabbed) 
 				this.grabbed = false;
 				
 		return super.mouseReleased(mouseX, mouseY, button);
 	}
-
 }
